Registered Services Page Called Error API for ISV user
